html not accepting floating point numbers
